Assenze orarie automaticamente giustificate se per uscite anticipate o ritardi. Eliminazione bug di cancellazione giustifiche su reinserimento lezione.

+++ lib/req_salva_assenze.php
@@ -8,12 +8,12 @@ $idal = $id['al'];
 $asslez = "oreass" . $idal;
 
 $asslezal = stringa_html($asslez);
-$esisteassenza=false;
+$esisteassenza = false;
 //print "id al: $idal $asslez $asslezal <br>";
 $query = "select * from tbl_asslezione where idalunno=$idal and idlezione=$idlezione";
-$ris= eseguiQuery($con, $query);
-if (mysqli_num_rows($ris)>0)
-    $esisteassenza=true;
+$ris = eseguiQuery($con, $query);
+if (mysqli_num_rows($ris) > 0)
+    $esisteassenza = true;
 
 if ($asslezal != 0) {
     $flaguscent = false;
@@ -33,24 +33,21 @@ if ($asslezal != 0) {
         }
     }
 
-    if (!$esisteassenza)
-    {
-    if (!$flaguscent)
-        $query = "insert into tbl_asslezione(idalunno, idlezione,idmateria,data,oreassenza,forzata) values ($idal,$idlezione,$idmateria,'$data',$asslezal,true)";
-    else
-        $query = "insert into tbl_asslezione(idalunno, idlezione,idmateria,data,oreassenza,forzata,giustifica,iddocentegiust,datagiustifica) values ($idal,$idlezione,$idmateria,'$data',$asslezal,true,true,0,'$data')";
-    }
-    else
+    if (!$esisteassenza) {
+        if (!$flaguscent)
+            $query = "insert into tbl_asslezione(idalunno, idlezione,idmateria,data,oreassenza,forzata) values ($idal,$idlezione,$idmateria,'$data',$asslezal,true)";
+        else
+            $query = "insert into tbl_asslezione(idalunno, idlezione,idmateria,data,oreassenza,forzata,giustifica,iddocentegiust,datagiustifica) values ($idal,$idlezione,$idmateria,'$data',$asslezal,true,true,0,'$data')";
+    } else
         $query = "update tbl_asslezione set  oreassenza=$asslezal where idalunno=$idal and idlezione=$idlezione";
-  
-    
+
+
     eseguiQuery($con, $query);
-    
-}
-else
-{
-    $query = "delete from tbl_asslezione where idalunno=$idal and idlezione=$idlezione";
-    eseguiQuery($con, $query);
+} else {
+    if ($esistelezione) {
+        $query = "delete from tbl_asslezione where idalunno=$idal and idlezione=$idlezione";
+        eseguiQuery($con, $query);
+    }
 }
